sql/sem: move AsJSON to package tree where it belongs
Release note: None
