+++ pkg/sql/sem/tree/datum_test.go
@@ -719,3 +719,13 @@ func TestIsDistinctFrom(t *testing.T) {
 		})
 	}
 }
+
+func TestAllTypesAsJSON(t *testing.T) {
+	for _, typ := range types.AnyNonArray {
+		d := tree.SampleDatum(typ)
+		_, err := tree.AsJSON(d)
+		if err != nil {
+			t.Errorf("couldn't convert %s to JSON: %s", d, err)
+		}
+	}
+}
